refactor: Added documentation for modules, moved prologue and epilogue statements to separate files
- There is quite a bit more reworking to do here, but I need to move
  this along. I will come back to this later.
